+++ scripts/UpdateUbuntu1404.sh
@@ -1,6 +1,8 @@
 #! /bin/bash
-source racrc
+# Set to same as image_name in the .json - a temporary name for building
+source ../rc_files/racrc
 
+cd ../images
 # Download the latest version
 wget -N https://cloud-images.ubuntu.com/daily/server/trusty/current/trusty-server-cloudimg-amd64-disk1.img
 
@@ -11,16 +13,23 @@ glance_id=`openstack image create --disk-format qcow2 --container-format bare --
 # Run Packer on RAC
 packer build \
     -var "source_image=$glance_id" \
-    Ubuntu1404.json
+    scripts/Ubuntu1404.json | tee ../logs/Ubuntu1404.log
+
+if [ ${PIPESTATUS[0]} != 0 ]; then
+    exit 1
+fi
 
 openstack image delete TempUbuntuImage
-
-echo "Image Available on RAC!"
+sleep 5
+#For some reason getting the ID fails but using the name succeeds
+openstack image set --property description="Built on `date`" --property image_type='image' "${IMAGE_NAME}"
 
 # Grab Image and Upload to DAIR
-openstack image save $glance_id --file 1404.img
+openstack image save ${IMAGE_NAME} --file 1404.img
+openstack image set --name "Ubuntu 14.04" "${IMAGE_NAME}"
+echo "Image Available on RAC!"
 
-source dairrc
+source ../rc_files/dairrc
 openstack image create --disk-format qcow2 --container-format bare --file 1404.img "Ubuntu 14.04"
 
 echo "Image Available on DAIR!"
